requirements.txt
asgiref==3.11.0
black==25.12.0
click==8.3.1
colorama==0.4.6
Django==6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
pytokens==0.3.0
sqlparse==0.5.4
tzdata==2025.3
stripe
